pyproject.toml
[project]
name = "weaver"
version = "0.1.0"
description = "Issue tracking for AI coding agents"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
weaver = "weaver.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/weaver"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
